requirements.txt
requests==2.31.0
pymongo==4.6.0
python-dotenv==1.0.0
tenacity==8.2.2
